moves files, fixes cli install

+++ deployments/docker-entrypoint.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
 set -e
-echo "installing isotope cli....."
 
-cargo install --path isotope-cli
+file="/isotope/isotope-cli"
+
+if [ -e "$file" ]; then
+	echo "isotope cli already installed"
+else
+	echo "installing isotope cli....."
+	cargo install --path isotope-cli
+fi
 
 exec "$@"
